requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
celery==5.3.4
redis==4.5.5
django-celery-beat==2.5.0
django-celery-results==2.5.1
python-dotenv==1.0.0
kombu==5.3.4
django_crontab==0.7.1